server/requirements.txt
asttokens==2.1.0
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.38.0
fpdf==1.7.2
googlemaps==4.7.0
h11==0.13.0
idna==3.3
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nltk==3.7
numpy==1.23.2
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.4
paramiko==2.12.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.33
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pymongo==4.2.0
PyNaCl==1.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
pytz==2022.2.1
regex==2022.10.31
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sshtunnel==0.4.0
stack-data==0.6.1
tqdm==4.64.0
traitlets==5.5.0
trio==0.21.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.12
wcwidth==0.2.5
webdriver-manager==3.8.3
websockets==10.3
Werkzeug==2.2.2
wordcloud==1.8.2.2
wsproto==1.2.0
XlsxWriter==3.0.3
